requirements.txt
Flask
Flask_Cors
groq
keras
matplotlib
numpy
Pillow
python-dotenv
requests
tensorflow